Moved TestUtil, added getHttpBinUrl and simplified HttpResponseDummy
