+++ src/main/java/de/dani09/http/HttpResponseDummy.kt
@@ -1,3 +1,3 @@
 package de.dani09.http
 
-class HttpResponseDummy @JvmOverloads constructor(responseCode: Int = 0) : HttpResponse(responseCode, ByteArray(0), mapOf())
+class HttpResponseDummy @JvmOverloads constructor(responseCode: Int = 0) : HttpResponse(responseCode, byteArrayOf(), mapOf())
